// src/utils/enums/TankTypes.ts
export enum TankTypes {
  PEER_MOVE = 0,
  TILE_PUNCH = 3,
  PEER_WORLD = 4,
  TILE_UPDATE = 5,
  PEER_ENTER_DOOR = 7,
  TILE_DAMAGE = 8,
  PEER_INVENTORY = 9,
  PEER_CLOTH = 10,
  PEER_COLLECT = 11,
  TILE_TREE = 12,
  PEER_DROP = 14,
  TILE_APPLY_LOCK = 15,
  PEER_ITEMS_DAT = 16,
  EFFECT = 17,
  PEER_ICON = 18,
  PEER_PING_RESPONSE = 21,
  PEER_PING_REQUEST = 22,
  PUNCH = 23,
}